move the util method to slice_util, and make the backend api match up with tf c signature
